Fully Implemented UserImp class that implements methods from my UserService class. Fixes #4. Fixes #3.
