+++ src/main/java/Interfaces/UserService.java
@@ -8,5 +8,6 @@ public interface UserService {
   public User getUser(Long id);
   User updateUser(User user);
   void deleteUser(Long id);
+  public User createUser(User user);
 
 }
