+++ src/main/java/repository/UserRepo.java
@@ -2,6 +2,8 @@ package repository;
 
 import entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface UserRepo extends JpaRepository<User, Long> {
 }
